Deleted unnecessary ECLIPSE files
