+++ CourseManagementGUI/resources/assignment-logo.png
