+++ CourseManagementGUI/resources/course-logo.png
