+++ CourseManagementGUI/resources/lms-logo.png
